==> project_workflow/1_data_collection.ipynb <==
# Moving raw data to processed data

# Data Collection and Processing
In this notebook, we: 
- collect a number of CSV and XLSX files
- clean, simplify, and reshape them
- merge them together by county

import numpy as np
import pandas as pd

## County Basics - All Counties in the US

counties = pd.read_csv("data/raw/uscounties.csv")
counties.head()

## County Census Data - US Census Bureau

age_education = pd.read_excel("data/raw/county_age_education_census.xlsx", sheet_name = "Data")
age_education.head(11)

age_education.columns

age_education.rename(columns = {"Unnamed: 0" : "county",
                      'Unnamed: 2' : "age_25+",
                      'Unnamed: 10' : "highschool_diploma",
                      'Unnamed: 11' : "college_degree"},
                    inplace = True)
education = age_education[["county",
                           "age_25+",
                           "highschool_diploma",
                           "college_degree"]].copy()

education.head(10)

# in this instance, we are choosing to disregard the margin of error metric from the US Census Bureau.
# We believe this is the best education data we can find.
education["county"] = education["county"].shift(2)
education.head(10)

education_notnull = education[education["county"].isnull() == False]
education_notnull[education_notnull["county"].str.contains("County")].reset_index()

age_education.shape
