==> project_workflow/3_cancer-water-data.ipynb <==
## Imports

import pandas as pd

## Read rivers data

df = pd.read_csv('data/cleaned/full_river_site_hab_chem_bminv.csv', index_col=[0])

df.shape

# df = df.drop(columns=['BANGMODE', 'XBKA', 'MEDBK_A', 'XUN', 'MEDBKUN', 'LRBS_BW5', 'LDCBF_G08', 'XCDENMID', 'XSHOR2VG', 
#                       'PCT_OVRB', 'CONPERCENT', 'LDMB_BW5','LRBS_G08','AG', 'LRBS_TST', 'SDWD_RAT', 'XCDENBK'])
# df = df.dropna(axis=1)

from IPython.display import display
with pd.option_context('display.max.columns', None):
    display(df.head(1))

## Read cancer data

cancer = pd.read_csv('data/cleaned/nih_cancer_counties_20221111.csv')

#df['County-State'] = (df['CNTYNAME'] + '-' + df['PSTL_CODE']).str.lower().head()

cancer['County-State'] = cancer['County-State'].str.replace(' county', '').str.replace(' parish', '')

## Merge river data to cancer data

cancer_water = pd.merge(df, cancer, on='County-State', how='outer')

cancer_water.isna().sum()

df.head()

county_list = zip(
    sorted(list(set(df['County-State']))), 
    sorted(list(set(cancer['County-State'])))
                 )

for i in county_list:
    print(i)

cancer.columns

cancer_water.dropna(inplace=True, axis=0)

cancer_water['is_rising'] = 0
cancer_water.loc[cancer_water['Recent Trend'] == 'rising', 'is_rising'] = 1

cancer_water.to_csv('data/cleaned/rivers-cancer-rates.csv')
