==> project_workflow/8_eda_shaping_epa_chems.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
pd.set_option("display.max_columns", 100)

cdr = pd.read_csv("data/raw/CDR_full_20221111.csv", index_col=[0])
cdr.head()

cdr.shape

cdr_reduced = cdr[["SITE_ZIP", "PARENT_COMPANY_NAME", "STRIPPED_CHEMICAL_ID_NUMBER", "CHEMICAL_NAME"]].copy()
cdr_reduced.columns = [col.lower() for col in cdr_reduced.columns]
cdr_reduced["site_zip"] = cdr_reduced["site_zip"].apply(lambda x : x.split("-")[0] if "-" in x else x)
cdr_reduced.replace("", np.nan, inplace = True)
# https://stackoverflow.com/questions/47333227/pandas-valueerror-cannot-convert-float-nan-to-integer
#cdr_reduced["site_zip"] = pd.to_numeric(cdr_reduced["site_zip"], errors = "coerce")
cdr_reduced["site_zip"] = cdr_reduced["site_zip"].astype(float)
#cdr_reduced["site_zip"].astype(float, errors = "ignore")
#cdr_reduced[cdr_reduced["site_zip"] == ""]


# https://www.kaggle.com/datasets/danofer/zipcodes-county-fips-crosswalk
zip_fips = pd.read_csv("data/raw/ZIP-COUNTY-FIPS_2017-06.csv")[["ZIP", "STCOUNTYFP"]]
zip_fips.columns = [col.lower() for col in zip_fips.columns]
zip_fips.head()                       

cdr_reduced = cdr_reduced.merge(zip_fips, left_on = "site_zip", right_on = "zip", how = "left")
cdr_reduced.head()

cdr_reduced.isnull().sum()

cdr_reduced.dropna(inplace = True)

cdr_reduced.head()

cdr_reduced = cdr_reduced[["stcountyfp", "zip", "parent_company_name", "stripped_chemical_id_number", "chemical_name"]].copy()
cdr_reduced["stcountyfp"] = cdr_reduced["stcountyfp"].astype(int)
cdr_reduced['stripped_chemical_id_number'] = cdr_reduced['stripped_chemical_id_number'].astype(int)
cdr_reduced.rename(columns = {"stcountyfp" : "fips"}, inplace = True)
cdr_reduced.head()

# count of CDR companies

cdr_reduced_companies_uniques = cdr_reduced.groupby("fips", as_index=False).nunique()
cdr_reduced_companies_uniques

# Create a dataframe of chemicals

cdr_uniques = cdr_reduced_companies_uniques[["fips", "parent_company_name", "stripped_chemical_id_number"]].copy()
cdr_uniques.rename(columns = {
    "parent_company_name" : "chem_companies",
    "stripped_chemical_id_number" : "unique_reported_chemicals"
}, inplace = True)
cdr_uniques.head()

cdr_reduced.head()

# fips to chemical ids
chems = cdr_reduced[["fips", "stripped_chemical_id_number"]].copy()

# the chemical names that we use later to change column names
col = 'stripped_chemical_id_number'
chemnames = cdr_reduced.assign(**{col: lambda x: x[col].astype(int)}).drop_duplicates([col]).set_index(col)['chemical_name']
chemnames.head()

chem_dummies = pd.get_dummies(chems, columns=["stripped_chemical_id_number"], drop_first=False).groupby('fips', as_index=False).sum()
chem_dummies.columns = [f"chem_{chemnames[int(c.rsplit('_', 1)[1])]}" if '_' in c else c for c in chem_dummies]

# merge the dummified chemical data to the unique company names and number of chemicals
print(chem_dummies.shape)
print(cdr_uniques.shape)
cdr_chemset = cdr_uniques.merge(chem_dummies, on="fips", how="left")
print(cdr_chemset.shape)

cdr_chemset.to_csv("data/processed/county_chemicals.csv", index = False)
