==> project_workflow/9_eda_shaping_county_statistics.ipynb <==
# County Statistics

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

counties_2019 = pd.read_csv("data/raw/county_2019.csv")
county_amer_com_surv = pd.read_csv("data/raw/county_info_openintro.csv")
counties_2019.head()

counties_2019.columns

county_amer_com_surv.head()

county_fips = counties_2019[["name", "fips", "age_over_65"]]
county_fips.head()

print(county_fips.shape)
print(county_amer_com_surv.shape)


np.mean(county_fips["name"] == county_amer_com_surv["name"]) #looks like these counties aren't quite exactly matched.

county_fips[county_fips["name"] != county_amer_com_surv["name"]]

county_amer_com_surv[county_amer_com_surv["name"] != county_fips["name"]]

# since these two county lists are identical except for punctiation, we will concat them rather than left join.
county_data = pd.concat([county_fips, county_amer_com_surv.drop(columns = "name")], axis = 1)
county_data

county_data.to_csv("data/processed/county_information.csv", index = False)
