==> project_workflow/15_MODEL_cancer-water-rf.ipynb <==
## Imports

import pandas as pd
import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier

## Read dataset

cancer_water = pd.read_csv('data/cleaned/rivers-cancer-rates.csv')

cancer_water = cancer_water.filter(regex='^(?!Unnamed).*')

## Model 1: Preprocessing

X = cancer_water.drop(columns=['is_rising', 'Age-Adjusted Incidence Rate([rate note]) - cases per 100,000',
       'Lower 95% Confidence Interval', 'Upper 95% Confidence Interval',
       'CI*Rank([rank note])', 'Lower CI (CI*Rank)', 'Upper CI (CI*Rank)',
       'Average Annual Count', 'Recent Trend',
       'Recent 5-Year Trend ([trend note]) in Incidence Rates',
       'Lower 95% Confidence Interval.1', 'Upper 95% Confidence Interval.1'])
y = cancer_water['is_rising']

cancer_water_dummies = pd.get_dummies(X, drop_first=True)

X_train, X_test, y_train, y_test = train_test_split(cancer_water_dummies, y, stratify=y, random_state=2022)

## Baseline

y.value_counts(normalize=True).max()

## Model 1: Random Forest

#X_dummies = pd.get_dummies(X_train, drop_first=True)

rf = RandomForestClassifier(n_estimators=200, n_jobs=-1)
rf.fit(X_train, y_train);

## Model 1: Evaluation

rf.score(X_train, y_train)

rf.score(X_test, y_test)

preds = rf.predict(X_train)

preds.sum()

preds = rf.predict(X_test)
preds.sum()

rf.feature_importances_
